✨ Support bulk affect updates at independent affects endpoint
This endpoint has not yet been implemented by OSIDB.

♻️ Split zod schemas up by entity

♻️ Split zod schemas up by entity

✨ Support bulk affect updates

🐛 Correct imports

♻️ Don't notify users when each affect is saved

🔥 Remove debugging statements
